(#1903) Moved dnt-entry to ftp-module + small compilation bugs
DNT-entry-creation  only used by ftp-module.
